Move header card parser to be hidden (private) static inner class of eaderCard. Nothing outside needs to be exposed to its inner workings.
